Add a simple network and a tutorial
